requirements.txt
discord.py>=2.0.0,<3.0.0
python-dotenv>=0.19.0,<1.0.0
pytest>=7.0.0,<8.0.0
black>=23.0.0,<24.0.0
isort>=5.0.0,<6.0.0
flake8>=6.0.0,<7.0.0
